bump sql v1.3

+++ NekoRobot/modules/sql/approve_sql.py
@@ -25,7 +25,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import threading
 
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, String, UnicodeText, BigInteger, func, distinct
 
 from NekoRobot.modules.sql import BASE, SESSION
 
@@ -33,7 +33,7 @@ from NekoRobot.modules.sql import BASE, SESSION
 class Approvals(BASE):
     __tablename__ = "approval"
     chat_id = Column(String(14), primary_key=True)
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
 
     def __init__(self, chat_id, user_id):
         self.chat_id = str(chat_id)  # ensure string
